QuickAdminPanel automatic commit
