+++ app/Grupo.php
@@ -18,14 +18,19 @@ class Grupo extends Model
     ];
 
     protected $fillable = [
-        'curso',
+        'curso_id',
         'created_at',
         'updated_at',
         'deleted_at',
     ];
 
-    public function horarios()
+    public function grupos()
     {
-        return $this->hasMany(Horario::class, 'curso_id', 'id');
+        return $this->hasMany(Grupo::class, 'curso_id', 'id');
+    }
+
+    public function curso()
+    {
+        return $this->belongsTo(Grupo::class, 'curso_id');
     }
 }
