+++ app/Http/Requests/StoreHorarioRequest.php
@@ -25,9 +25,8 @@ class StoreHorarioRequest extends FormRequest
             'dia'        => [
                 'required',
             ],
-            'curso_id'   => [
+            'curso'      => [
                 'required',
-                'integer',
             ],
             'asignatura' => [
                 'required',
