+++ resources/lang/es/cruds.php
@@ -347,10 +347,10 @@ return [
             'updated_at_helper' => '',
             'deleted_at'        => 'Deleted at',
             'deleted_at_helper' => '',
-            'curso'             => 'Curso',
-            'curso_helper'      => 'Seleccione un curso',
             'color'             => 'Color',
             'color_helper'      => 'Seleccione un color de los que estan abajo.',
+            'curso'             => 'Curso',
+            'curso_helper'      => 'Introduzca el curso. Ej: ESO 1ºA, BAC 1ºA, . . . .',
         ],
     ],
     'grupo'           => [
