+++ resources/views/admin/grupos/edit.blade.php
@@ -10,17 +10,18 @@
         <form action="{{ route("admin.grupos.update", [$grupo->id]) }}" method="POST" enctype="multipart/form-data">
             @csrf
             @method('PUT')
-            <div class="form-group {{ $errors->has('curso') ? 'has-error' : '' }}">
+            <div class="form-group {{ $errors->has('curso_id') ? 'has-error' : '' }}">
                 <label for="curso">{{ trans('cruds.grupo.fields.curso') }}*</label>
-                <input type="text" id="curso" name="curso" class="form-control" value="{{ old('curso', isset($grupo) ? $grupo->curso : '') }}" required>
-                @if($errors->has('curso'))
+                <select name="curso_id" id="curso" class="form-control select2" required>
+                    @foreach($cursos as $id => $curso)
+                        <option value="{{ $id }}" {{ (isset($grupo) && $grupo->curso ? $grupo->curso->id : old('curso_id')) == $id ? 'selected' : '' }}>{{ $curso }}</option>
+                    @endforeach
+                </select>
+                @if($errors->has('curso_id'))
                     <p class="help-block">
-                        {{ $errors->first('curso') }}
+                        {{ $errors->first('curso_id') }}
                     </p>
                 @endif
-                <p class="helper-block">
-                    {{ trans('cruds.grupo.fields.curso_helper') }}
-                </p>
             </div>
             <div>
                 <input class="btn btn-danger" type="submit" value="{{ trans('global.save') }}">
